refactor: #17 reorganize packages for user auth
