--- services/identity-service/src/main/java/io/ciphernance/identity/application/command/user/authenticate/AuthenticateUserCommand.java
@@ -1,10 +1,0 @@
-package io.ciphernance.identity.application.command.user.authenticate;
-
-import io.ciphernance.identity.application.mediator.Command;
-
-public record AuthenticateUserCommand(
-
-        String email,
-        String password,
-        String totpCode
-) implements Command<AuthenticateUserResponse> { }
