+++ services/identity-service/src/main/java/io/ciphernance/identity/application/command/user/authenticate/register/RegisterUserCommand.java
@@ -1,4 +1,4 @@
-package io.ciphernance.identity.application.command.user.register;
+package io.ciphernance.identity.application.command.user.authenticate.register;
 
 import io.ciphernance.identity.application.mediator.Command;
